citation_processor/requirements.txt
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
pytest==7.4.3 